More stuff

+++ src/main/java/io/stubbs/selenium/SeleniumUtils.java
@@ -1,11 +1,14 @@
 package io.stubbs.selenium;
 
 import com.google.common.base.Stopwatch;
+import io.github.bonigarcia.wdm.WebDriverManager;
+import io.stubbs.captcha.FunctionUtils;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
 
 import javax.imageio.ImageIO;
 import java.awt.*;
@@ -15,14 +18,33 @@ import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.util.Base64;
 
+import static io.stubbs.captcha.FunctionUtils.logStopwatch;
+import static io.stubbs.captcha.FunctionUtils.time;
 import static org.openqa.selenium.OutputType.BYTES;
-import static org.openqa.selenium.OutputType.FILE;
 
 @Slf4j
 @RequiredArgsConstructor
 public class SeleniumUtils {
 
     final private ChromeDriver driver;
+
+    private boolean deleteOnExit = false;
+
+    @SneakyThrows
+    static public ChromeDriver getChromeDriver() {
+        WebDriverManager.chromedriver().setup();
+        ChromeOptions chromeOptions = new ChromeOptions();
+
+//        File tempFile = File.createTempFile("touch", "txt", null);
+//        String absolutePath = tempFile.getPath() + "/ocado/touch";
+//        new File(absolutePath).mkdirs();
+
+//        chromeOptions.addArguments("user-data-dir=" + "/private/var/folders/wl/3xwk5p0x22q76b_yrfkpb1pw0000gp/T/ocado");
+
+        ChromeDriver driver = new ChromeDriver(chromeOptions);
+
+        return driver;
+    }
 
     @SneakyThrows
     public String shotElement(WebElement ele) {
@@ -34,7 +56,7 @@ public class SeleniumUtils {
         // String screenshotAs = driver.getScreenshotAs(BASE64);
         Stopwatch w = Stopwatch.createStarted();
         byte[] screenshot = driver.getScreenshotAs(BYTES);
-        log.info(w.elapsed().toString());
+        logStopwatch(log, w);
         BufferedImage fullImg = ImageIO.read(new ByteArrayInputStream(screenshot));
         log.info(w.elapsed().toString());
 //        log.info("Read screenshot off disk...");
@@ -60,10 +82,10 @@ public class SeleniumUtils {
                 scaledWidth,
                 scaledHeight,
                 BufferedImage.TYPE_INT_RGB);
-        log.info("Scaling image...");
-        log.info(w.elapsed().toString());
-        scaledImage.createGraphics().drawImage(eleScreenshot, 0, 0, scaledWidth, scaledHeight, Color.WHITE, null);
-        log.info(w.elapsed().toString());
+
+        Graphics2D graphics = time("Scaling image", () -> scaledImage.createGraphics());
+
+        time("Drawing", () -> graphics.drawImage(eleScreenshot, 0, 0, scaledWidth, scaledHeight, Color.WHITE, null));
 
 
         // Copy the element screenshot to disk
@@ -71,12 +93,12 @@ public class SeleniumUtils {
         // tmpFile.deleteOnExit();
         // File screenshotLocation = new File("C:\\images\\GoogleLogo_screenshot.png");
         File screenshotLocationjpg = File.createTempFile("screenshot-part-jpg", ".jpg");
-        screenshotLocationjpg.deleteOnExit();
+        if (deleteOnExit)
+            screenshotLocationjpg.deleteOnExit();
 
         // File screenshotLocation = File.createTempFile("screenshot-part", ".png");
 
         log.info("Saving scaled image to disk...");
-        log.info(w.elapsed().toString());
         boolean jpg1 = ImageIO.write(scaledImage, "jpg", screenshotLocationjpg);
         if (!jpg1) throw new RuntimeException("Failed to find writer...");
 
@@ -87,15 +109,12 @@ public class SeleniumUtils {
         // BufferedOutputStream output = new BufferedOutputStream(scaledOutputBuffer);
         log.info("Save ...");
         boolean jpg = ImageIO.write(scaledImage, "jpg", scaledOutputBuffer);
-        log.info(w.elapsed().toString());
         if (!jpg) throw new RuntimeException("Failed to find writer...");
 
         scaledOutputBuffer.close();
 
         log.info("Base64 encoding...");
-        log.info(w.elapsed().toString());
         String base64EncodedImage = Base64.getEncoder().encodeToString(scaledOutputBuffer.toByteArray());
-        log.info(w.elapsed().toString());
         // FileUtils.copyFile(screenshot, screenshotLocation);
         return base64EncodedImage;
     }
